plugin monte carlo test is split up to three tests
