+++ test/plugin_monte_carlo_bs/PluginMonteCarloTests.cpp
@@ -1,0 +1,5 @@
+#include "PluginMonteCarloTests.h"
+
+PluginMonteCarloTests::PluginMonteCarloTests() {
+    pluginsModelsDir = rrTestModelsDir_ / "PLUGINS";
+}
